requirements.txt
aiohttp==3.4.4
aiohttp-jinja2==1.0.0
python-socketio==2.0.0
